Fix CouchDB to cross-compile on the Mac
Patch the configure script instead of replacing it with a generated one as the version of libtool may not be the same on all hosts.
Add check in the installer script that a file exists before updating paths in it.
Fix icu-config script to work with the configure script in case "echo -n" doesn't work.